third_party/WebKit/LayoutTests/platform/mac/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/redirect-loop-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: http://127.0.0.1:8000/fetch/resources/redirect-loop.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&Count=20&mode=cors&credentials=same-origin&method=GET should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
